refactor: refactor expense management code
