+++ client/src/utils/index.js
@@ -5,6 +5,7 @@ export { default as objUtils } from './objUtils';
 export { default as authFieldsCheck } from './authFieldsCheck';
 export { default as stringUtils } from './stringUtils';
 export { default as colorConverters } from './colorConverters';
+export { default as money } from './money';
 
 export function cn(...inputs) {
   return twMerge(clsx(inputs));
